menu.go: ensure that owner-draw/bitmap separators are flagged as such
The code that flags separators was being treated as mutually-exculsive
to owner-draw/bitmap items. We move the separator stuff out of the
switch block into its own if statement to ensure that it is always run
regardless of an item's owner-draw/bitmap setting.

Updates https://github.com/tailscale/tailscale/issues/17535

Signed-off-by: Aaron Klotz <aaron@tailscale.com>